TDS_API/pyproject.toml
[tool.poetry]
name = "tds-api"
version = "0.1.0"
description = "API REST TABLE DES SAVEURS"
authors = ["mama"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
ipython = "^8.26.0"
sentry-sdk = "^2.13.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
